# server/docs/swagger.yaml
definitions:
...
  request.BatchUpdateStatusReq:
    properties:
      ids:
        items:
          type: integer
        type: array
      status:
        type: integer
    type: object
...
  request.GetById:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.PageInfo:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  system.SysDataSource:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      database:
        description: 数据库名称
        type: string
      host:
        description: 主机地址
        type: string
      name:
        description: 数据源名称
        type: string
      password:
        description: 密码
        type: string
      port:
        description: 端口
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态 0禁用 1启用
        type: integer
      type:
        description: MySQL/达梦/人大金仓/PostgreSQL/Oracle/SQLServer
        type: string
      updatedAt:
        description: 更新时间
        type: string
      username:
        description: 用户名
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.1
paths:
...
  /dataSource/batchDelete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除数据源
      tags:
      - DataSource
  /dataSource/batchUpdateStatus:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID列表和状态
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BatchUpdateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量更新数据源状态
      tags:
      - DataSource
  /dataSource/createDataSource:
    post:
      consumes:
      - application/json
      parameters:
      - description: 数据源信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysDataSource'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建数据源
      tags:
      - DataSource
  /dataSource/deleteDataSource:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除数据源
      tags:
      - DataSource
  /dataSource/getDataSourceById:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.SysDataSource'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取数据源详情
      tags:
      - DataSource
  /dataSource/getDataSourceList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取数据源列表
      tags:
      - DataSource
  /dataSource/getTables:
    get:
      consumes:
      - application/json
      parameters:
      - description: 数据源ID
        in: query
        name: dataSourceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 表列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取数据源下的表列表
      tags:
      - DataSource
  /dataSource/testConnection:
    post:
      consumes:
      - application/json
      parameters:
      - description: 数据源信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysDataSource'
      produces:
      - application/json
      responses:
        "200":
          description: 测试结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 测试数据源连接
      tags:
      - DataSource
  /dataSource/updateDataSource:
    post:
      consumes:
      - application/json
      parameters:
      - description: 数据源信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysDataSource'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新数据源
      tags:
      - DataSource
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...